+++ source/renderer/zettlr-rendereripc.js
@@ -1,5 +1,3 @@
-/* global $ */
-
 /**
 * @ignore
 * BEGIN HEADER