+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - The text snippets have been removed from the app.
 - The Quicklook windows now follow the application's theme.
 - Fixed a bug that generated a falsy first search cursor and prevented case insensitive searching.
+- Added extended statistics. Now you can exactly see when you've written how many words by clicking the new button in the small statistics popup.
 
 ## Under the hood